moved pages to pages folder and created two new pages
