+++ src/App.jsx
@@ -1,5 +1,5 @@
 import './App.css'
-import Home from './components/Home'
+import Home from './pages/Home'
 import Menu from './components/Menu'
 
 function App() {
