+++ src/pages/About.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const About = () => {
+  return (
+    <div id='about-content'>
+        <p>Här kommer det vara lite info</p>
+    </div>
+  )
+}
+
+export default About
