+++ src/pages/Projects.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const Projects = () => {
+  return (
+    <div>Projects</div>
+  )
+}
